requirements.txt
# TTR Signal Detection System Requirements

# Core dependencies
PyYAML>=6.0
requests>=2.31.0
python-dotenv>=1.0.0

# Development dependencies (install with: pip install -e .[dev])
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
